method_installer removes the blessed state of subs being imported (RT#103247)
also see https://github.com/rjbs/Sub-Exporter-ForMethods/issues/1

I cannot reproduce this on my own machine, so I do not know what version
combinations are responsible. For now, the safest thing is to simply avoid
installing subs in this way, and skip checking this namespace's cleanliness.

The one failure report for this I've seen is:
http://www.cpantesters.org/cpan/report/af1ea6a0-a028-1014-88d0-c10f2b825c07